+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/tasks/GenerateParserTask.kt
@@ -48,11 +48,15 @@ abstract class GenerateParserTask : JavaExec() {
 
     /**
      * The location of the generated parser class, relative to the [targetRootOutputDir].
-     * The value of this property is only used when [purgeOldFiles] is set to `true`.
+     * For backwards compatibility, setting this property changes the default of [purgeOldFiles] to `false`.
+     * Otherwise, the value of this property is only used when [purgeOldFiles] is set to `true`.
      * When this property is set, only the given file or directory is deleted before re-generating the parser.
      */
     @Deprecated(
         message = "The property is removed without replacement. " +
+                "Be aware that the task may delete all files in `targetRootOutputDir` after unsetting this property. " +
+                "You may set `purgeOldFiles = false` to avoid deleting the files. " +
+                "When this property and `pathToPsiRoot` is not set, `purgeOldFiles` gets enabled by default. " +
                 "This property is only relevant when the directory at `targetRootOutputDir` overlaps with other files. " +
                 "Note that overlapping task outputs are discouraged by Gradle and can cause issues with the build cache.",
         level = DeprecationLevel.WARNING,
@@ -63,11 +67,15 @@ abstract class GenerateParserTask : JavaExec() {
 
     /**
      * The location of the generated PSI files, relative to the [targetRootOutputDir].
-     * The value of this property is only used when [purgeOldFiles] is set to `true`.
+     * For backwards compatibility, setting this property changes the default of [purgeOldFiles] to `false`.
+     * Otherwise, the value of this property is only used when [purgeOldFiles] is set to `true`.
      * When this property is set, only the given file or directory is deleted before re-generating the parser.
      */
     @Deprecated(
         message = "The property is removed without replacement. " +
+                "Be aware that the task may delete all files in `targetRootOutputDir` after unsetting this property. " +
+                "You may set `purgeOldFiles = false` to avoid deleting the files. " +
+                "When this property and `pathToParser` is not set, `purgeOldFiles` gets enabled by default. " +
                 "This property is only relevant when the directory at `targetRootOutputDir` overlaps with other files. " +
                 "Note that overlapping task outputs are discouraged by Gradle and can cause issues with the build cache.",
         level = DeprecationLevel.WARNING,
@@ -98,6 +106,11 @@ abstract class GenerateParserTask : JavaExec() {
 
     /**
      * Purge old files from the target directory before generating the parser.
+     * By default, old files are purged unless [pathToParser] and [pathToPsiRoot] have been specified.
+     * If you want to disable this option because the output directory is shared with another task,
+     * note that you may run into issues with stale files. Also note that
+     * [overlapping task outputs are discouraged by Gradle](https://docs.gradle.org/current/userguide/organizing_gradle_projects.html#avoid_overlapping_task_outputs)
+     * and may cause issues when using the build cache.
      */
     @get:Input
     @get:Optional
@@ -110,10 +123,12 @@ abstract class GenerateParserTask : JavaExec() {
         } else if (pathToPsiRoot.isPresent && !pathToParser.isPresent) {
             throw GradleException("'pathToPsiRoot' has a configured value, but 'pathToParser' has not. You must either remove or keep both properties.")
         }
-        if (purgeOldFiles.orNull == true) {
-            if (!pathToParser.isPresent && !pathToPsiRoot.isPresent) {
+        purgeOldFiles.orNull.also { purge ->
+            if (purge == false) {
+                // Do nothing as `purgeOldFiles` is explicitly disabled
+            } else if (!pathToParser.isPresent && !pathToPsiRoot.isPresent) {
                 targetRootOutputDir.get().asFile.deleteRecursively()
-            } else {
+            } else if (purge == true) {
                 // Delete only the directories specified by `pathToParser` and `pathToPsiRoot` for backwards compatibility.
                 targetRootOutputDir.get().asFile.apply {
                     resolve(pathToParser.get()).deleteRecursively()