+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Deprecate property `targetOutputDir` of `GenerateLexerTask`.
 - Deprecate properties `pathToParser` and `pathToPsiRoot` of `GenerateParserTask`.
+- Purge stale files `GenerateLexerTask` and `GenerateParserTask` by default (as soon as you stop using the deprecated properties).
 - Deprecate `targetFile(String)` and `targetFile(Provider)` of `GenerateLexerTask`.
 - Deprecate `parserFile()` and `psiDir()` of `GenerateParserTask`.